requirements.txt
pytesseract
opencv-python
mysql-connector-python
Pillow
